feat: 문제 종류에 따라서 푼 문제 저장하는 로직 구체화

+++ backend/user/src/main/java/com/scf/user/profile/domain/entity/Solved.java
@@ -32,7 +32,7 @@ public class Solved {
     private String choice;
 
     @Column(name = "problem_id")
-    private int problemId;
+    private Long problemId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id") // FK 설정
